6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ long int c;
 long long int d;
 float f;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned int long)sizeof(a));
-printf("Size of a int: %lu byte(s)\n", (unsigned int long)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n", (unsigned int long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned int long)sizeof(d));
-printf("Size of a float: %lu byte(s)\n", (unsigned int long)sizeof(f));
+printf("Size of a char: %zu byte(s)\n", sizeof(a));
+printf("Size of a int: %zu byte(s)\n", sizeof(b));
+printf("Size of a long int: %zu byte(s)\n", sizeof(c));
+printf("Size of a long long int: %zu byte(s)\n", sizeof(d));
+printf("Size of a float: %zu byte(s)\n", sizeof(f));
 return (0);
 }
